修复打包崩溃
