+++ webview/views/viewProblem/submissionTypes.ts
@@ -1,0 +1,10 @@
+export type ProblemSubmissionContext = {
+  fileName?: string;
+  filePath?: string;
+  languages: {
+    label: string;
+    id: number;
+    O2?: true;
+  }[];
+  defaultLanguage?: string;
+};
